requirements.txt
# Requerimientos básicos
requests>=2.28.2
pymysql>=1.0.3
python-dotenv>=1.0.0
colorama>=0.4.6
schedule>=1.2.0
cryptography>=39.0.2
Jinja2>=3.1.2
Werkzeug>=2.3.7

# API de Google
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.80.0

# Framework web y autenticación
Flask>=2.3.3
Flask-Login>=0.6.2
Flask-WTF>=1.1.1
WTForms>=3.0.1
email-validator>=2.0.0

# Seguridad
bcrypt>=4.0.1
Pillow>=9.5.0
itsdangerous>=2.1.2

# Utilidades
pytz>=2023.3
python-dateutil>=2.8.2